azure-mcp-server/requirements.txt
azure-functions==1.20.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-monitor-opentelemetry==1.6.1
msal==1.30.0
pyjwt[crypto]==2.9.0
cryptography==42.0.8
aiohttp==3.10.5
sse-starlette==2.1.3
pydantic==2.8.2
python-jose==3.3.0
opencensus-ext-azure==1.1.13
asyncio==3.4.3
python-multipart==0.0.9